requirements.txt
torch
torchvision
numpy
pandas
matplotlib
plotly
seaborn
tqdm
dataset
Pillow
ipykernel
albumentations
model
